Rename spatialAudio to privateZones
